Delete product: an owner or an admin can delete products from the website. If other users had the product in favorite they will delete from the list.

+++ ByteMarket/backend/src/main/java/es/grupo04/backend/service/ProductService.java
@@ -21,12 +21,16 @@ import es.grupo04.backend.model.Purchase;
 import es.grupo04.backend.model.User;
 import es.grupo04.backend.repository.ProductRepository;
 import es.grupo04.backend.repository.PurchaseRepository;
+import es.grupo04.backend.repository.UserRepository;
 
 @Service
 public class ProductService {
 
 	@Autowired
 	private ProductRepository repository;
+
+	@Autowired
+	private UserRepository userRepository;
 
 	@Autowired
 	private PurchaseRepository purchaseRepository;
@@ -46,25 +50,37 @@ public class ProductService {
 	public List<Product> findAll() {
 		return repository.findAll();
 	}
-
+	// Save a product
 	public Product save(Product product) {
 		return repository.save(product);
 	}
-
+	// Delete a product
 	public void delete(long id) {
-		repository.deleteById(id);
+		Optional<Product> productOpt = repository.findById(id);
+        if (productOpt.isPresent()) {
+            Product product = productOpt.get();
+            //Look for users with the product as a favorite
+            List<User> usersWithFavoriteProduct = repository.findUsersByFavoriteProduct(product);
+            //Eliminate the product from the favorites of the users
+            for (User user : usersWithFavoriteProduct) {
+                user.getFavoriteProducts().remove(product);	
+                userRepository.save(user);
+            }
+            //Eliminate the product from the owner's products
+            repository.deleteById(id);
+        }
 	}
-
+	// Get the last purchases of a users
     public List<Purchase> getLastPurchases(User user) {
         return purchaseRepository.findByBuyerOrderByPurchaseDateDesc(user).stream()
                 .limit(5)
                 .collect(Collectors.toList());
     }
-
+	// Get the last sales of a user
 	public List<Purchase> getLastSales(User user) {
         return purchaseRepository.findBySellerOrderByPurchaseDateDesc(user);
     }
-
+	// Add images to a product
 	public void addImages(Product product, MultipartFile[] images) throws IOException {
 		ArrayList<Image> imagesToStore = new ArrayList<>();
         for(MultipartFile image : images) {
@@ -76,7 +92,7 @@ public class ProductService {
 		product.setImages(imagesToStore);
 		product.setThumbnail(imagesToStore.get(0));
     }
-
+	// Pagination 
 	public List<Product> findPaginated(int page, int pageSize) {
 		Pageable pageable = PageRequest.of(page, pageSize);
 		Page<Product> productPage = repository.findAll(pageable);
